[otbn] Teach RIG to generate BEQ and BNE instructions
The branching program flow means that we have to teach the RIG some
new tricks. Firstly, we need to add "recursive" generation (so that
the generator for a branch can generate stuff in each side below it).
We do this by passing a continuation of type GenCont to the generator.

Secondly, we need to tighten up the snippet type a bit. Branches are
represented as branches in the expression tree (as you'd expect), and
we also needed to add a sequencing type, SeqSnippet (so that you can
have a list of instructions followed by a branch, for example).

We always generate branches that converge (so there will only ever be
one ECALL in the generated program). To glue the branches back
together, we generate a jump from the end of one to the end of the
other. This needs slight changes in the Jump generator so that we can
specify the target address that we want in this case.

The final big change is that we need to add support for "φ nodes" to
the model, to cope with when execution joins back together. This is
the "merge" method added to Model, CallStack and KnownMem.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>