[aes] Use sparse encodings for FSMs, trigger alert for invalid states
This commit changes all FSMs inside AES to use sparse state encodings with
a minimum Hamming distance of 3 between states. If any FSM enters an
invalid state, the AES unit locks and an alert is triggered. The cipher
core is prevented from finishing the current operation or starting a new
operation. Writes to the control register are ignored. The AES unit needs
to be reset.

This commit also renames the alert signals. Update errors in the shadow
control register are collected in `recoverable`. More severe errors such as
storage errors in the shadow control register, entering invalid FSM
states trigger a `fatal` alert.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>