[clkmgr / top] clean-up core file usage
There was some duplicated core file usage in clkmgr due to the ip / template split.

This duplicated usage was causing top_earlgrey_asic lint to break.

For now, remove the ip level core files completely because there is no plan for there to be a clkmgr block level specific DV, nor is there anyone other than top that references clkmgr_pkg.

If this changes, we need to move back to the ip level re-direct.

Signed-off-by: Timothy Chen <timothytim@google.com>