[dv/sram] add SRAM scrambling model for DV
This PR adds a DV golden model for the scrambling mechanisms used by the
SRAM to scramble both addresses and data, into the `mem_bkdr_if`.

The main feature is a custom substitution/permutation network similar to
the PRESENT block cipher, which is added to the `sram_ctrl_env_pkg`.

Data scrambling uses the PRINCE block cipher as well, so the existing
DPI model is imported and re-used here.

An API for performing encrypted memory reads and writes is also added to
the interface so users can easily perform encrypted memory accesses.

This PR also updates the parameter names and values in the SRAM
testbench to match the updated `mem_bkdr_if`.

This PR also adds parity support and basic ECC support when performing
writes or otherewise modifying the memory.

Signed-off-by: Udi Jonnalagadda <udij@google.com>