[otbn/dif] Remove logging from smoke test
The smoke test did log a fair amount of status information during the
execution. Doing so is rather slow on Verilator simulations (due to the
way the simulated UART works). Remove the logging to speed up test
execution in CI, at the cost of making the debugging of this test
slightly harder. Logging in error cases is still the same as before.

Also removed is information on how long the applications took to execute on
OTBN. This was useful information in early bringup, but isn't that
useful any more for the smoke test.

Fixes #4742

Signed-off-by: Philipp Wagner <phw@lowrisc.org>