[doc] Add package install command for CentOS 7
This commit adds instructions for installing package dependencies for
CentOS 7, which is among the most used Linux distributions in the hardware
world. At the moment there are only a few subtle differences between the
dependencies for Ubuntu 18.04 and CentOS 7, but this might change in
the future.

This is related to lowRISC/OpenTitan#3299.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>