[DV] Enable cov comp creation iff cov is enabled
In our DV base classes (`hw/dv/sv/dv_lib`), the `en_cov` cfg bit is set
to 1 by default, for both, the agents and the env. The original
motivation for this was to enable coverage-driven stimulus generation.
This is something we are not doing at the moment. Even if we decide to
do something like that, it can be enabled on a per test / vseq basis.

This change sets the `en_cov` bit at the env cfg level to 0 and enables
this setting to be retrieved via plusarg in `dv_base_test`. In all env
classes, this cfg setting is then passed on the the downstream agent cfg
objects. This change is applied to `uvmdvgen` env template as well.

In `hw/dv/tools/dvsim/common_modes.hjson` the `+en_cov=1` plusarg is
set, so that if coverage collection is enabled, the coverage collection
component is created (previously, the coverage component was created
regardless of whether the coverage was enabled or not).

With this change, simulations run without coverage collection enabled
should see roughly 10-30% improvement in runtime, which directly impacts
the private CI runtime (which has been inching closer to the 20 min
timeout as more DUTs are added to the regression list).

Signed-off-by: Srikrishna Iyer <sriyer@google.com>