[secded_gen] Add ability to generate FPV TB's and correct Hamming code
This adds additional templates to automatically generate a simple FPV
testbench for each SECDEC pair.

Also, this slightly reworks the encoder/decoder templates to be more
compact, and corrects the Hamming variant (the parity bit did not
include all other ECC bits, and the error status decode was not
correct). The Hamming code errors have been flagged by the FPV testbench
(to be checked in in the subsequent commit).

Signed-off-by: Michael Schaffner <msf@opentitan.org>