[otbn] Allow multiple errors in a single cycle in ISS
This means we can't just raise an exception any more, which makes
things a little more fiddly. Instead, any component of the state that
might raise an error has an "errs" list, to which it appends an error
if it sees one. These lists get concatenated in a tree, in much the
same way as we gather up changes.

After executing an instruction, the code checks whether there were any
errors. If not, it commits pending changes. If there is an error, it
figures out the correct value of ERR_BITS and stops.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>