[dv/otp_ctrl] regwen sequence
This PR adds a regwen sequence in OTP_CTRL. The timing to check
direct_access_regwen is very sensitive because it is written by HW and
write to 0 only when DAI operation is in-progress.
This PR follows keymgr's cfgen sequence example and read
direct_access_regwen register during DAI busy, and write one of the
enabled registers when access is locked.

Signed-off-by: Cindy Chen <chencindy@google.com>