[tlul] Wire up sync fifo full_o output port to tlul modules
This output port was added in commit 71d98f8e4. It's roughly the same
as the inverse of the wready_o signal and tlul_adapter_sram,
tlul_fifo_sync don't care about the difference between the two.

Also, reorder the ports to match the order in prim_fifo_sync.sv to
make things a bit easier to read.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>