[ci] Quote argument to get-build-type.sh
When we run on master (rather than a PR),
$SYSTEM_PULLREQUEST_TARGETBRANCH isn't defined, since this isn't a
pull request. Make sure we still actually pass 2 arguments to
get-build-type.sh.

Also, condition the Python lint check on being a PR. Amusingly, this
was running on the master branch before, but the call to git to get a
file list was broken (which we didn't notice), hiding any follow-on
errors.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>