[clkgen] Replace IBUFG with IBUF
The IBUFG primitive doesn't exist in 7 series+ devices and is replaced
by a normal IBUF during syntheis, as Vivado notes:

```
[Netlist 29-432] The IBUFG primitive 'clkgen/io_clk_ibufg' has been
retargeted to an IBUF primitive only. No BUFG will be added. If a
global buffer is intended, please instantiate an available global
clock primitive from the current architecture.
```

In our case we don't need a global buffer as the output of the IBUF(G)
primitive will be fed into a PLL block before being used for logic,
and we add appropriate global clock buffers there.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>