[ci] Remove "wrong" build artifacts from SW builds
We build the device software twice: once in a normal version for
top_earlgrey, and once in a patched version which is only used/suitable
for the Nexys Video FPGA board. The whole patching process is a rather
ugly hack, but necessary until we are able to build multiple parametrizable
toplevels, e.g. with reduced amounts of flash and SRAM to match
available resources on certain FPGA boards.

Whenever we build software, the build outputs end up in $BIN_DIR. If we
build the software twice from different sources (i.e., the unpatched and
the patched sources), we have build artifacts with the same file name,
but different content in $BIN_DIR. That makes the whole situation not
only extremely confusing, but also dangerously buggy: if we download build
artifacts from both the unpatched and the patched build step and extract
them into the same directory (as we do currently), we get an
unpredictable mix of binaries from different sources.

This patch addresses the worst part of the problem by removing all build
artifacts from $BIN_DIR which we know shouldn't be there.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>