[ci] Do not upload artifacts from GCC SW build
We build the device software once with GCC and once with Clang. Only the
Clang-built software is (meant to be) used, the GCC build is only done
as a sanity check to keep GCC working. Its outputs should not end up
being used by subsequent steps, or end up in the distribution archive.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>